Rely on auto_increment value
In theory, we could explicitly insert values into the id column that do
not update the auto_increment_value (e.g. negative numbers). But in
practice, this is never done with Active Record.

Therefore, if a table has an auto_increment_value, we rely on its value.
Only for tables without auto_increment columns we have to look for rows.